--- src/Util/index.ts
@@ -1,1 +1,0 @@
-export * from './Validate';
